+++ src/app/api/upload/types.ts
@@ -1,0 +1,20 @@
+
+export const allowedMimeTypes = [
+  "image/jpeg",
+  "image/png",
+  "image/gif",
+  "image/webp",
+  "image/svg+xml",
+  "image/bmp",
+  "image/tiff",
+] as const;
+
+export const mimeTypeToFileExtension: Record<(typeof allowedMimeTypes)[number], string> = {
+  "image/jpeg": ".jpg",
+  "image/png": ".png",
+  "image/gif": ".gif",
+  "image/webp": ".webp",
+  "image/svg+xml": ".svg",
+  "image/bmp": ".bmp",
+  "image/tiff": ".tiff",
+} as const;
